refactor: move IHasher and IHashCompare to use cases interface folder
