+++ src/use-cases/interfaces/hash-compare-interface.ts
