+++ src/use-cases/interfaces/hasher-interface.ts
