+++ src/use-cases/interfaces/index.ts
@@ -2,3 +2,5 @@ export * from '@/use-cases/interfaces/use-case-interface';
 export * from '@/use-cases/interfaces/universally-unique-identifier-generator-interface';
 export * from '@/use-cases/interfaces/email-options-interface';
 export * from '@/use-cases/interfaces/email-service-interface';
+export * from '@/use-cases/interfaces/hasher-interface';
+export * from '@/use-cases/interfaces/hash-compare-interface';
