+++ src/use-cases/user/interfaces/index.ts
@@ -3,8 +3,6 @@ export * from '@/use-cases/user/interfaces/user-editable-data-interface';
 export * from '@/use-cases/user/interfaces/user-repository-interface';
 export * from '@/use-cases/user/interfaces/user-repository-return-data-interface';
 export * from '@/use-cases/user/interfaces/user-repository-data-interface';
-export * from '@/use-cases/user/interfaces/hasher-interface';
-export * from '@/use-cases/user/interfaces/hash-compare-interface';
 export * from '@/use-cases/user/interfaces/encrypter-interface';
 export * from '@/use-cases/user/interfaces/decrypter-interface';
 export * from '@/use-cases/user/interfaces/payload-interface';
